lab1/server/requirements.txt
alembic==1.12.1
asyncpg==0.29.0
email-validator==2.1.0.post1
fastapi==0.104.1
passlib==1.7.4
pre-commit==3.5.0
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-jose==3.3.0
sqlalchemy==2.0.23
uvicorn==0.23.2
python-dotenv==1.0.0
